Fix toggle pane code to be aligned with code styles
