+++ code/src/UI/Controls/TogglePane.cs
@@ -13,72 +13,105 @@
 using System.Windows;
 using System.Windows.Controls;
 
+using Microsoft.Templates.UI.Extensions;
+
 namespace Microsoft.Templates.UI.Controls
 {
-    public sealed partial class TogglePane
+    public sealed partial class TogglePane : Control
     {
-        #region TogglePaneContent
+        private Border _togglePaneShadowGrid;
+        private Grid _menuGrid;
+        private bool _isInitialized = false;
+
         public object TogglePaneContent
         {
-            get { return GetValue(TogglePaneContentProperty); }
-            set { SetValue(TogglePaneContentProperty, value); }
+            get => GetValue(TogglePaneContentProperty);
+            set => SetValue(TogglePaneContentProperty, value);
         }
         public static readonly DependencyProperty TogglePaneContentProperty = DependencyProperty.Register("TogglePaneContent", typeof(object), typeof(TogglePane), new PropertyMetadata(null));
-        #endregion
 
-        #region MainViewTemplate
         public DataTemplate MainViewTemplate
         {
-            get { return (DataTemplate)GetValue(MainViewTemplateProperty); }
-            set { SetValue(MainViewTemplateProperty, value); }
+            get => (DataTemplate)GetValue(MainViewTemplateProperty);
+            set => SetValue(MainViewTemplateProperty, value);
         }
         public static readonly DependencyProperty MainViewTemplateProperty = DependencyProperty.Register("MainViewTemplate", typeof(DataTemplate), typeof(TogglePane), new PropertyMetadata(null));
-        #endregion
 
-        #region MainViewTemplateSelector
         public DataTemplateSelector MainViewTemplateSelector
         {
-            get { return (DataTemplateSelector)GetValue(MainViewTemplateSelectorProperty); }
-            set { SetValue(MainViewTemplateSelectorProperty, value); }
+            get => (DataTemplateSelector)GetValue(MainViewTemplateSelectorProperty);
+            set => SetValue(MainViewTemplateSelectorProperty, value);
         }
         public static readonly DependencyProperty MainViewTemplateSelectorProperty = DependencyProperty.Register("MainViewTemplateSelector", typeof(DataTemplateSelector), typeof(TogglePane), new PropertyMetadata(null));
-        #endregion
 
-        #region SecondaryViewTemplate
         public DataTemplate SecondaryViewTemplate
         {
-            get { return (DataTemplate)GetValue(SecondaryViewTemplateProperty); }
-            set { SetValue(SecondaryViewTemplateProperty, value); }
+            get => (DataTemplate)GetValue(SecondaryViewTemplateProperty);
+            set => SetValue(SecondaryViewTemplateProperty, value);
         }
         public static readonly DependencyProperty SecondaryViewTemplateProperty = DependencyProperty.Register("SecondaryViewTemplate", typeof(DataTemplate), typeof(TogglePane), new PropertyMetadata(null));
-        #endregion
 
-        #region OpenButtonTemplate
         public DataTemplate OpenButtonTemplate
         {
-            get { return (DataTemplate)GetValue(OpenButtonTemplateProperty); }
-            set { SetValue(OpenButtonTemplateProperty, value); }
+            get => (DataTemplate)GetValue(OpenButtonTemplateProperty);
+            set => SetValue(OpenButtonTemplateProperty, value);
         }
         public static readonly DependencyProperty OpenButtonTemplateProperty = DependencyProperty.Register("OpenButtonTemplate", typeof(DataTemplate), typeof(TogglePane), new PropertyMetadata(null));
-        #endregion
 
-        #region CloseButtonTemplate
         public DataTemplate CloseButtonTemplate
         {
-            get { return (DataTemplate)GetValue(CloseButtonTemplateProperty); }
-            set { SetValue(CloseButtonTemplateProperty, value); }
+            get => (DataTemplate)GetValue(CloseButtonTemplateProperty);
+            set => SetValue(CloseButtonTemplateProperty, value);
         }
         public static readonly DependencyProperty CloseButtonTemplateProperty = DependencyProperty.Register("CloseButtonTemplate", typeof(DataTemplate), typeof(TogglePane), new PropertyMetadata(null));
-        #endregion
 
-        #region IsOpen
         public bool IsOpen
         {
-            get { return (bool)GetValue(IsOpenProperty); }
-            set { SetValue(IsOpenProperty, value); }
+            get => (bool)GetValue(IsOpenProperty);
+            set => SetValue(IsOpenProperty, value);
+        }
+        public static readonly DependencyProperty IsOpenProperty = DependencyProperty.Register("IsOpen", typeof(bool), typeof(TogglePane), new PropertyMetadata(false, OnIsOpenPropertyChanged));
+
+        public bool AllowDragAndDrop
+        {
+            get => (bool)GetValue(AllowDragAndDropProperty);
+            set => SetValue(AllowDragAndDropProperty, value);
+        }
+        public static readonly DependencyProperty AllowDragAndDropProperty = DependencyProperty.Register("AllowDragAndDrop", typeof(bool), typeof(TogglePane), new PropertyMetadata(false));
+
+        static TogglePane()
+        {
+            DefaultStyleKeyProperty.OverrideMetadata(typeof(TogglePane), new FrameworkPropertyMetadata(typeof(TogglePane)));
         }
 
-        public static readonly DependencyProperty IsOpenProperty = DependencyProperty.Register("IsOpen", typeof(bool), typeof(TogglePane), new PropertyMetadata(false, OnIsOpenPropertyChanged));
+        public override void OnApplyTemplate()
+        {
+            base.OnApplyTemplate();
+            _togglePaneShadowGrid = GetTemplateChild("togglePaneShadowGrid") as Border;
+            _menuGrid = GetTemplateChild("menuGrid") as Grid;
+
+            _isInitialized = true;
+            UpdateOpenStatus();
+        }
+
+        private void UpdateOpenStatus(bool newValue = false, bool oldValue = false)
+        {
+            if (_isInitialized)
+            {
+                if (IsOpen)
+                {
+                    _menuGrid.AnimateWidth(90);
+                    _togglePaneShadowGrid.AnimateWidth(90);
+                    _togglePaneShadowGrid.FadeIn();
+                }
+                else
+                {
+                    _menuGrid.AnimateWidth(30);
+                    _togglePaneShadowGrid.AnimateWidth(30);
+                    _togglePaneShadowGrid.FadeOut();
+                }
+            }
+        }
 
         private static void OnIsOpenPropertyChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
         {
@@ -88,16 +121,5 @@ namespace Microsoft.Templates.UI.Controls
                 control.UpdateOpenStatus((bool)e.NewValue, (bool)e.OldValue);
             }
         }
-        #endregion
-
-        #region AllowDragAndDrop
-        public bool AllowDragAndDrop
-        {
-            get { return (bool)GetValue(AllowDragAndDropProperty); }
-            set { SetValue(AllowDragAndDropProperty, value); }
-        }
-
-        public static readonly DependencyProperty AllowDragAndDropProperty = DependencyProperty.Register("AllowDragAndDrop", typeof(bool), typeof(TogglePane), new PropertyMetadata(false));
-        #endregion
     }
 }
