+++ code/src/UI/Controls/TogglePane.xaml
