+++ code/src/UI/Themes/Generic.xaml
@@ -3,6 +3,6 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
     <ResourceDictionary.MergedDictionaries>
-        <ResourceDictionary Source="/Microsoft.Templates.UI;component/Controls/TogglePane/TogglePane.xaml" />
+        <ResourceDictionary Source="/Microsoft.Templates.UI;component/Controls/TogglePane.xaml" />
     </ResourceDictionary.MergedDictionaries>
 </ResourceDictionary>
